allow Machine to hook into foreign statics; remove unused HasMemory trait
